# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: eQuadras API - Gestão e Agendamento Esportivo
  description: "Documentação oficial das APIs REST da plataforma eQuadras para integraç\
    ões de sistemas parceiros, bots e clientes de API. Todos os endpoints incluem\
    \ exemplos reais de requisição e resposta."
  contact:
    name: Suporte Técnico eQuadras
    url: https://github.com/GuilhermeAizzaSano/eQuadras
    email: contato@equadras.app
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: v1.0.0
servers:
- url: https://equadras.app
  description: Servidor de Produção (HTTPS)
- url: /
  description: Servidor Relativo (Mesma Origem)
- url: http://localhost:8080
  description: Ambiente Local (Desenvolvimento)
security:
- BearerAuth: []
- ApiKeyAuth: []
...
paths:
...
  /api/quadras/{quadraId}/desbloquear:
    post:
      tags:
      - Bloqueios - API
      summary: Desbloquear horários ou dia (Admin)
      description: "Papel: ADMIN dono da quadra ou Master Admin. Informe `bloqueioId`\
        \ OU `data` (com `horaInicio`/`horaFim` opcionais para desbloquear só um intervalo).\
        \ Sem nenhum dos dois, 400. Devolve `{mensagem, totalRemovidos}` com quantos\
        \ bloqueios foram removidos."
      operationId: desbloquear
      parameters:
      - name: quadraId
        in: path
        description: ID da quadra (`id_quadra`)
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DesbloqueioHorarioDTO"
            example:
              bloqueioId: 5
              data: 2026-10-12
              horaInicio: 14:00:00
              horaFim: 16:00:00
        required: true
      responses:
        "200":
          description: Resultado do desbloqueio
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
              example:
                mensagem: Horários desbloqueados com sucesso.
                totalRemovidos: 1
        "400":
          description: Requisição Inválida
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/bad-request
                title: Requisição Inválida
                status: 400
                detail: Informe o ID do bloqueio ou a data a ser desbloqueada.
                instance: "/api/quadras/{quadraId}/desbloquear"
                code: REQUISICAO_INVALIDA
                timestamp: 2026-09-29T14:30:00.123456Z
        "403":
          description: Acesso Negado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/acesso-negado
                title: Acesso Negado
                status: 403
                detail: Apenas o administrador dono da quadra ou o Master Admin pode
                  remover bloqueios.
                instance: "/api/quadras/{quadraId}/desbloquear"
                code: ACESSO_NEGADO
                timestamp: 2026-09-29T14:30:00.123456Z
...
  /api/quadras/{id}/bloqueios:
    get:
      tags:
      - Bloqueios - API
      summary: Listar bloqueios da quadra
      description: "Papéis: CLIENT ou ADMIN. Lista os bloqueios ativos e futuros de\
        \ uma quadra. Quadra inexistente devolve 400."
      operationId: listarBloqueios
      parameters:
      - name: id
        in: path
        description: ID da quadra (`id_quadra`)
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Bloqueios ativos e futuros da quadra
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BloqueioHorarioResponseDTO"
              example:
              - id: 5
                quadraId: 1
                data: 2026-10-12
                horaInicio: 14:00:00
                horaFim: 18:00:00
                motivo: Torneio Interno da Arena
                criadoEm: 2026-09-29T09:30:00
        "400":
          description: Requisição Inválida
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/bad-request
                title: Requisição Inválida
                status: 400
                detail: A quadra informada não foi encontrada.
                instance: "/api/quadras/{id}/bloqueios"
                code: REQUISICAO_INVALIDA
                timestamp: 2026-09-29T14:30:00.123456Z
    post:
      tags:
      - Bloqueios - API
      summary: Criar bloqueio de horário ou dia inteiro (Admin)
      description: "Papel: ADMIN dono da quadra ou Master Admin. Sem `horaInicio`/`horaFim`\
        \ bloqueia o dia inteiro; com horário, ambos são obrigatórios e o início precisa\
        \ ser anterior ao fim. A data não pode estar no passado. Falha com 400 se\
        \ já houver reserva ativa no período ou bloqueio coincidente; se o dia já\
        \ está totalmente bloqueado, a mensagem começa com `DIA_INTEIRO_BLOQUEADO`\
        \ e `substituirDiaInteiro=true` desbloqueia o restante do dia mantendo só\
        \ o novo intervalo."
      operationId: criarBloqueio
      parameters:
      - name: id
        in: path
        description: ID da quadra (`id_quadra`)
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloqueioHorarioCriacaoDTO"
            example:
              data: 2026-10-12
              horaInicio: 14:00:00
              horaFim: 18:00:00
              motivo: Torneio Interno da Arena
              substituirDiaInteiro: false
        required: true
      responses:
        "201":
          description: Bloqueio criado
          content:
            application/json:
              example:
                id: 5
                quadraId: 1
                data: 2026-10-12
                horaInicio: 14:00:00
                horaFim: 18:00:00
                motivo: Torneio Interno da Arena
                criadoEm: 2026-09-29T09:30:00
        "400":
          description: Requisição Inválida
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              examples:
                REQUISICAO_INVALIDA:
                  summary: Já existe um bloqueio cadastrado que coincide com este
                    horário nesta data.
                  value:
                    type: https://api.equadras.com/erros/bad-request
                    title: Requisição Inválida
                    status: 400
                    detail: Já existe um bloqueio cadastrado que coincide com este
                      horário nesta data.
                    instance: "/api/quadras/{id}/bloqueios"
                    code: REQUISICAO_INVALIDA
                    timestamp: 2026-09-29T14:30:00.123456Z
                REQUISICAO_INVALIDA_2:
                  summary: Não é possível bloquear este horário pois já existem reservas
                    ativas no período.
                  value:
                    type: https://api.equadras.com/erros/bad-request
                    title: Requisição Inválida
                    status: 400
                    detail: Não é possível bloquear este horário pois já existem reservas
                      ativas no período.
                    instance: "/api/quadras/{id}/bloqueios"
                    code: REQUISICAO_INVALIDA
                    timestamp: 2026-09-29T14:30:00.123456Z
        "422":
          description: Dados Inválidos
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/validacao
                title: Dados Inválidos
                status: 422
                detail: "Dados inválidos: data: A data do bloqueio é obrigatória"
                instance: "/api/quadras/{id}/bloqueios"
                code: ERRO_VALIDACAO
                timestamp: 2026-09-29T14:30:00.123456Z
                camposIncorretos:
                - campo: Campo
                  mensagem: Mensagem de validação do campo
        "403":
          description: Acesso Negado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/acesso-negado
                title: Acesso Negado
                status: 403
                detail: Apenas o administrador dono da quadra ou o Master Admin pode
                  criar bloqueios.
                instance: "/api/quadras/{id}/bloqueios"
                code: ACESSO_NEGADO
                timestamp: 2026-09-29T14:30:00.123456Z
  /api/quadras:
    get:
      tags:
      - Quadras - API
      summary: "Listar quadras ativas, por proximidade e filtros"
      description: |-
        Papéis: CLIENT ou ADMIN. Lista quadras ativas com filtros opcionais (texto parcial, sem diferenciar maiúsculas) e busca por proximidade (`latitude` + `longitude` + `raioKm`).

        **Formato da resposta** (decidido nesta ordem, a primeira regra que se aplica vence):
        1. `resumido=true|false` explícito;
        2. header `X-Client: frontend` ou `X-View: full` → completo;
        3. header `X-View: resumo|summary` ou `X-Client: api` → resumido;
        4. sem nada disso: URL com `/api/` → resumido (padrão de bots e integrações).

        **Variantes:** (a) resumido → `array<QuadraResumoResponseDTO>`; `page` e `size` são ignorados; (b) completo com `page` → `PageQuadraResponseDTO` (size padrão 6, máximo 50); (c) completo sem `page` → `array<QuadraResponseDTO>`.

        **Teto:** listas sem paginação (a e c) devolvem no máximo 200 itens, limite aplicado no SQL e sem aviso ao cliente: use `page` (formato completo) ou filtros para não perder resultados.
      operationId: listarTodas
      parameters:
      - name: latitude
        in: query
        description: "Latitude do ponto de busca por proximidade (graus, -90 a 90).\
          \ Só filtra se `longitude` também vier."
        required: false
        schema:
          type: number
          format: double
        example: -20.8113
      - name: longitude
        in: query
        description: "Longitude do ponto de busca por proximidade (graus, -180 a 180).\
          \ Só filtra se `latitude` também vier."
        required: false
        schema:
          type: number
          format: double
        example: -49.3758
      - name: raioKm
        in: query
        description: "Raio da busca por proximidade, em km. Usado apenas com `latitude`\
          \ e `longitude`."
        required: false
        schema:
          type: string
        example: 2.0
      - name: tipoEsporte
        in: query
        description: "Filtra pelo tipo de esporte. Valores: FUTEBOL, FUTSAL, VOLEI,\
          \ BEACH_TENNIS, BASQUETE, TENIS."
        required: false
        schema:
          type: string
        example: FUTEBOL
      - name: nome
        in: query
        description: Filtra por parte do nome da quadra.
        required: false
        schema:
          type: string
        example: Arena
      - name: endereco
        in: query
        description: Filtra por parte do endereço (logradouro).
        required: false
        schema:
          type: string
        example: Av. Brasil
      - name: cidade
        in: query
        description: Filtra pela cidade.
        required: false
        schema:
          type: string
        example: São José do Rio Preto
      - name: bairro
        in: query
        description: Filtra pelo bairro.
        required: false
        schema:
          type: string
        example: Jardim das Flores
      - name: cep
        in: query
        description: Filtra pelo CEP.
        required: false
        schema:
          type: string
        example: 15000-000
      - name: resumido
        in: query
        description: "Força o formato: `true` = resumido (`QuadraResumoResponseDTO`),\
          \ `false` = completo (`QuadraResponseDTO`). Quando informado, tem precedê\
          ncia sobre os headers `X-Client`/`X-View` e sobre a URL."
        required: false
        schema:
          type: boolean
        example: true
      - name: page
        in: query
        description: "Índice da página, começando em 0. Só tem efeito no formato completo\
          \ e o transforma em resposta paginada (`PageQuadraResponseDTO`). IGNORADO\
          \ no formato resumido."
        required: false
        schema:
          type: integer
          format: int32
        example: 0
      - name: size
        in: query
        description: Itens por página. Padrão 6; máximo 50 (valores maiores são limitados
          a 50; zero ou negativo volta para 6). Só tem efeito no formato completo
          com `page`. IGNORADO no formato resumido.
        required: false
        schema:
          type: string
          maximum: 50
        example: 6
      - name: X-Client
        in: header
        description: Identifica o cliente. `frontend` força o formato completo; `api`
          força o resumido (só vale se `resumido` não vier).
        required: false
        schema:
          type: string
          enum:
          - frontend
          - api
        example: api
      - name: X-View
        in: header
        description: "Formato desejado. `full` = completo; `resumo` ou `summary` =\
          \ resumido (só vale se `resumido` não vier). Se `X-Client` também vier,\
          \ `X-Client: frontend` e `X-View: full` são avaliados antes de `X-View:\
          \ resumo|summary` e `X-Client: api`."
        required: false
        schema:
          type: string
          enum:
          - full
          - resumo
          - summary
        example: resumo
      responses:
        "200":
          description: "Formato resumido: `array<QuadraResumoResponseDTO>`. É o padrã\
            o em /api/quadras. `page` e `size` são ignorados e a lista é limitada\
            \ a 200 itens."
          content:
            application/json:
              schema:
                oneOf:
                - type: array
                  items:
                    $ref: "#/components/schemas/QuadraResumoResponseDTO"
                - $ref: "#/components/schemas/PageQuadraResponseDTO"
                - type: array
                  items:
                    $ref: "#/components/schemas/QuadraResponseDTO"
              examples:
                resumido:
                  summary: "Formato resumido: `array<QuadraResumoResponseDTO>`. É\
                    \ o padrão em /api/quadras. `page` e `size` são ignorados e a\
                    \ lista é limitada a 200 itens."
                  value:
                  - id_quadra: 1
                    nome: Arena Gol Society
                    tipoEsporte: FUTEBOL
                    valorHora: 120.0
                    endereco: "Av. Brasil, 1500, Jardim das Flores - São José do Rio\
                      \ Preto"
                    cep: 15000-000
                completoPaginado:
                  summary: "Formato completo com `page`: `PageQuadraResponseDTO` (size\
                    \ padrão 6, máximo 50)."
                  value:
                    content:
                    - id_quadra: 1
                      nome: Arena Gol Society
                      tipoEsporte: FUTEBOL
                      valorHora: 120.0
                      ativa: true
                      cep: 15000-000
                      logradouro: "Av. Brasil, 1500"
                      bairro: Jardim das Flores
                      cidade: São José do Rio Preto
                      estado: SP
                      latitude: -20.8113
                      longitude: -49.3758
                      descricao: Grama sintética padrão FIFA com iluminação em LED
                        e vestiários.
                      dataLimiteAgendamento: 2026-12-31
                      fotos:
                      - /uploads/quadras/3f2a9c1e-7b4d-4e8a-9c21-5d6f7e8a9b0c.jpg
                      disponibilidades:
                      - diaSemana: MONDAY
                        horaInicio: 08:00:00
                        horaFim: 22:00:00
                      versao: 3
                    pageable:
                      pageNumber: 0
                      pageSize: 6
                      sort:
                        empty: false
                        sorted: true
                        unsorted: false
                      offset: 0
                      paged: true
                      unpaged: false
                    last: true
                    totalPages: 1
                    totalElements: 1
                    size: 6
                    number: 0
                    sort:
                      empty: false
                      sorted: true
                      unsorted: false
                    first: true
                    numberOfElements: 1
                    empty: false
                completoLista:
                  summary: "Formato completo sem `page`: `array<QuadraResponseDTO>`,\
                    \ limitada a 200 itens sem aviso."
                  value:
                  - id_quadra: 1
                    nome: Arena Gol Society
                    tipoEsporte: FUTEBOL
                    valorHora: 120.0
                    ativa: true
                    cep: 15000-000
                    logradouro: "Av. Brasil, 1500"
                    bairro: Jardim das Flores
                    cidade: São José do Rio Preto
                    estado: SP
                    latitude: -20.8113
                    longitude: -49.3758
                    descricao: Grama sintética padrão FIFA com iluminação em LED e
                      vestiários.
                    dataLimiteAgendamento: 2026-12-31
                    fotos:
                    - /uploads/quadras/3f2a9c1e-7b4d-4e8a-9c21-5d6f7e8a9b0c.jpg
                    disponibilidades:
                    - diaSemana: MONDAY
                      horaInicio: 08:00:00
                      horaFim: 22:00:00
                    versao: 3
        "400":
          description: Parâmetro Inválido
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/parametro-invalido
                title: Parâmetro Inválido
                status: 400
                detail: "O parâmetro 'raioKm' possui um valor inválido: 'abc'."
                instance: /api/quadras
                code: PARAMETRO_INVALIDO
                timestamp: 2026-09-29T14:30:00.123456Z
        "401":
          description: Não Autorizado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                status: 401
                title: Não Autorizado
                detail: "Acesso não autorizado: credencial ausente, expirada ou invá\
                  lida."
    post:
      tags:
      - Quadras - API
      summary: Cadastrar nova quadra (Admin)
      description: "Papel: ADMIN (a quadra passa a pertencer ao administrador autenticado).\
        \ Cria a quadra com valor/hora, localização, `dataLimiteAgendamento` opcional\
        \ (não pode estar no passado), até 5 URLs de foto e até 7 regras de disponibilidade\
        \ semanal (um dia da semana só pode aparecer uma vez). Idempotente por `Idempotency-Key`:\
        \ reenvio com a mesma chave em até 10 minutos devolve a quadra já criada em\
        \ vez de duplicar. O campo `versao` do corpo é ignorado no POST."
      operationId: cadastrar_1
      parameters:
      - name: Idempotency-Key
        in: header
        description: "Chave gerada pelo cliente a cada formulário (ex.: UUID). Reenvios\
          \ com a mesma chave em até 10 minutos devolvem a quadra já criada. Opcional:\
          \ sem ela, cada chamada cria uma quadra."
        required: false
        schema:
          type: string
        example: 7c9e6679-7425-40de-944b-e07fc1f90ae7
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuadraCriacaoDTO"
            example:
              nome: Arena Gol Society
              tipoEsporte: FUTEBOL
              valorHora: 120.0
              cep: 15000-000
              logradouro: "Av. Brasil, 1500"
              bairro: Jardim das Flores
              cidade: São José do Rio Preto
              estado: SP
              latitude: -20.8113
              longitude: -49.3758
              descricao: Grama sintética padrão FIFA com iluminação em LED e vestiários.
              dataLimiteAgendamento: 2026-12-31
              fotos: []
              disponibilidades:
              - diaSemana: MONDAY
                horaInicio: 08:00:00
                horaFim: 22:00:00
        required: true
      responses:
        "201":
          description: Quadra criada. Reenvio com o mesmo `Idempotency-Key` em até
            10 minutos devolve a quadra já criada.
          content:
            application/json:
              example:
                id_quadra: 1
                nome: Arena Gol Society
                tipoEsporte: FUTEBOL
                valorHora: 120.0
                ativa: true
                cep: 15000-000
                logradouro: "Av. Brasil, 1500"
                bairro: Jardim das Flores
                cidade: São José do Rio Preto
                estado: SP
                latitude: -20.8113
                longitude: -49.3758
                descricao: Grama sintética padrão FIFA com iluminação em LED e vestiários.
                dataLimiteAgendamento: 2026-12-31
                fotos:
                - /uploads/quadras/3f2a9c1e-7b4d-4e8a-9c21-5d6f7e8a9b0c.jpg
                disponibilidades:
                - diaSemana: MONDAY
                  horaInicio: 08:00:00
                  horaFim: 22:00:00
                versao: 3
        "422":
          description: Dados Inválidos
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/validacao
                title: Dados Inválidos
                status: 422
                detail: "Dados inválidos: Nome: O nome deve ter entre 3 e 100 caracteres"
                instance: /api/quadras
                code: ERRO_VALIDACAO
                timestamp: 2026-09-29T14:30:00.123456Z
                camposIncorretos:
                - campo: Campo
                  mensagem: Mensagem de validação do campo
        "400":
          description: Requisição Inválida
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/bad-request
                title: Requisição Inválida
                status: 400
                detail: Uma quadra pode ter no máximo 5 fotos.
                instance: /api/quadras
                code: REQUISICAO_INVALIDA
                timestamp: 2026-09-29T14:30:00.123456Z
        "403":
          description: Acesso Proibido
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                status: 403
                title: Acesso Proibido
                detail: "Acesso proibido: sua credencial não possui permissão para\
                  \ executar esta operação."
        "401":
          description: Não Autorizado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                status: 401
                title: Não Autorizado
                detail: "Acesso não autorizado: credencial ausente, expirada ou invá\
                  lida."
...
  /api/quadras/{quadraId}/bloqueios/{bloqueioId}:
    delete:
      tags:
      - Bloqueios - API
      summary: Remover bloqueio por ID (Admin)
      description: "Papel: ADMIN dono da quadra ou Master Admin. Remove um bloqueio.\
        \ O bloqueio precisa pertencer à quadra do path (senão 400). Responde 204\
        \ sem corpo."
      operationId: removerBloqueio
      parameters:
      - name: quadraId
        in: path
        description: ID da quadra (`id_quadra`)
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: bloqueioId
        in: path
        description: ID do bloqueio (`id` de `BloqueioHorarioResponseDTO`)
        required: true
        schema:
          type: integer
          format: int64
        example: 5
      responses:
        "204":
          description: Bloqueio removido
        "400":
          description: Requisição Inválida
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/bad-request
                title: Requisição Inválida
                status: 400
                detail: O bloqueio informado não pertence a esta quadra.
                instance: "/api/quadras/{quadraId}/bloqueios/{bloqueioId}"
                code: REQUISICAO_INVALIDA
                timestamp: 2026-09-29T14:30:00.123456Z
        "403":
          description: Acesso Negado
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemaErro"
              example:
                type: https://api.equadras.com/erros/acesso-negado
                title: Acesso Negado
                status: 403
                detail: Apenas o administrador dono da quadra ou o Master Admin pode
                  remover bloqueios.
                instance: "/api/quadras/{quadraId}/bloqueios/{bloqueioId}"
                code: ACESSO_NEGADO
                timestamp: 2026-09-29T14:30:00.123456Z
...
components:
  schemas:
    LocalTime:
      type: string
      description: Horário no formato HH:mm:ss
      example: 14:00:00
...
    DesbloqueioHorarioDTO:
      type: object
      description: Dados para desbloquear horários de uma quadra via API
      properties:
        bloqueioId:
          type: integer
          format: int64
          description: ID específico do bloqueio cadastrado (opcional se informada
            a data)
          example: 1
        data:
          type: string
          format: date
          description: Data do bloqueio a ser removido (obrigatório se bloqueioId
            não informado)
          example: 2026-09-05
        horaFim:
          $ref: "#/components/schemas/LocalTime"
          description: Hora de término do bloqueio (opcional)
          example: 16:00:00
        horaInicio:
          $ref: "#/components/schemas/LocalTime"
          description: Hora de início do bloqueio (opcional)
          example: 14:00:00
...
    BloqueioHorarioResponseDTO:
      type: object
      properties:
        criadoEm:
          type: string
          format: date-time
          description: Data e hora de criação do bloqueio
          example: 2026-09-29T09:30:00
        data:
          type: string
          format: date
          description: Data do bloqueio
          example: 2026-10-12
        horaFim:
          $ref: "#/components/schemas/LocalTime"
          description: Fim do bloqueio
          example: 18:00:00
        horaInicio:
          $ref: "#/components/schemas/LocalTime"
          description: Início do bloqueio
          example: 14:00:00
        id:
          type: integer
          format: int64
          description: ID do bloqueio
          example: 5
        motivo:
          type: string
          description: Motivo informado na criação
          example: Torneio Interno da Arena
        quadraId:
          type: integer
          format: int64
          description: ID da quadra bloqueada
          example: 1
...
    ProblemaErro:
      type: object
      description: "Corpo de erro (RFC 9457). Os filtros de segurança devolvem só\
        \ status, title e detail."
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        timestamp:
          type: string
        camposIncorretos:
          type: array
          items:
            type: object
            properties:
              campo:
                type: string
              mensagem:
                type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      description: "Informe o token JWT ou API-Key no formato: `Bearer eq_...` gerado\
        \ no painel da conta."
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      description: Chave de API pessoal gerada no painel do usuário (formato `eq_...`).
        Válida em todas as rotas da API respeitando os papéis (ROLE_CLIENT e ROLE_ADMIN).
      name: X-API-KEY
      in: header